Separate search and shoebox modules and their dependencies
